file/get_info uses long instead of int for file size

+++ src/main/java/com/mediafire/sdk/api_responses/file/GetInfoResponse.java
@@ -114,11 +114,11 @@ public class GetInfoResponse extends ApiResponse {
             return this.description;
         }
 
-        public int getSize() {
+        public long getSize() {
             if (this.size == null) {
                 this.size = "0";
             }
-            return Integer.valueOf(this.size);
+            return Long.valueOf(this.size);
         }
 
         public boolean isPrivate() {
